Updates to force user confirmation for destructive operation

+++ .mongoshrc.js
@@ -211,13 +211,23 @@ usage.dropCollections =
     nsPattern - regex/string to limit collections/namespaces dropped
   Returns:
     { ok: ..., err: <error>, results: [ { db: <db>, cols: [ { col: <col>, dropped: <result> } ] } ] }
-  WARNING: THIS DOES NOT PROMPT FOR CONFIRMATION.`;
 
-function dropCollections(nsPattern) {
-  var ret = { ok: 1, results: [] }
+  NOTICE: This operation will prompt for user confirmation, enter 'YES' to confirm
+          OR pass 'silent' option as the last parameter to bypass confirmation i.e:
+          dropCollections(nsPattern, silent)`;
+
+function dropCollections(nsPattern, Confirmation = getUserConfirmation) {
+  var ret = { ok: 1, results: [] };
+  var cya = Confirmation;
+
   try {
     // Use getCollections() to confirm which collections will be dropped based upon the nsPattern
-    getCollections(nsPattern).results.forEach(function(dbObjIn) {
+    const results = getCollections(nsPattern).results;
+
+    if(!cya("drop", "collections", results))
+      throw `User must confirm this action ... dropCollections aborted!`;
+
+    results.forEach(function(dbObjIn) {
       var db = dbObjIn.db;
       var dbObjOut = { db: db, cols: [] };
       dbObjIn.cols.forEach(function(col) {
@@ -245,13 +255,23 @@ usage.dropDatabases =
     pattern - regex/string to limit databases dropped
   Returns:
     { ok: ..., err: <error>, results: [ { ok: 1, dropped: <db> } ] }
-  WARNING: THIS DOES NOT PROMPT FOR CONFIRMATION.`;
 
-function dropDatabases(pattern) {
+  NOTICE: This operation will prompt for user confirmation, enter 'YES' to confirm
+          OR pass 'silent' option as the last parameter to bypass confirmation i.e:
+          dropDatabases(pattern, silent)`;
+
+function dropDatabases(pattern, Confirmation = getUserConfirmation) {
   var ret = { ok: 1, results: [] };
+  var cya = Confirmation;
+
   try {
     // Use getDatabases() to confirm which databases will be dropped based upon the nsPattern
-    getDatabases(pattern).results.forEach(function(database) {
+    const results = getDatabases(pattern).results;
+
+    if(!cya("drop", "databases", results))
+      throw `User must confirm this action ... dropDatabases aborted!`;
+
+    results.forEach(function(database) {
       ret.results.push(getDatabase(database).dropDatabase());
     });
   } catch (error) {
@@ -274,14 +294,23 @@ usage.dropIndexes =
      idxPattern - regex/string to indexes
    Returns:
      { ok: ..., err: <error>, results: { ns: <namespace>, result: [ { nIndexesWas: ..., ok: ..., "$clusterTime": { clusterTime: ..., signature: ..., keyId: ...  } }, operationTime: ...  } ] }
-   WARNING: THIS DOES NOT PROMPT FOR CONFIRMATION.`;
 
-function dropIndexes(nsPattern, idxPattern) {
-  ret = { ok: 1 }
-  ret.results = [];
+  NOTICE: This operation will prompt for user confirmation, enter 'YES' to confirm
+          OR pass 'silent' option as the last parameter to bypass confirmation i.e:
+          dropIndexes(nsPattern, idxPattern, silent)`;
+
+function dropIndexes(nsPattern, idxPattern, Confirmation = getUserConfirmation) {
+  var ret = { ok: 1, results: [] };
+  var cya = Confirmation;
+
   try {
-    getIndexes(nsPattern, idxPattern).results.forEach(function(nsObj) {
-      var dbObj = { "ns": nsObj.ns, results: [] }
+    const results = getIndexes(nsPattern, idxPattern).results;
+
+    if(!cya("drop", "indexes", results))
+      throw `User must confirm this action ... dropIndexes aborted!`;
+
+    results.forEach(function(nsObj) {
+      var dbObj = { "ns": nsObj.ns, results: [] };
       nsObj.indexes.forEach(function(idx) {
         if(idx.name !== '_id_') {
           dbObj.results.push(getCollection(nsObj.ns).dropIndex(idx.name));
@@ -1249,3 +1278,59 @@ function getStableAPIStatus() {
   });
   return ret;
 }
+
+function getUserConfirmation(action, targetType, targetList) {
+  var userInput = 'NO';
+
+  print( `The following matching ${targetType} were found: ` );
+  print('');
+
+  switch (targetType) {
+    case 'databases':
+      targetList.forEach(function (dbObj) {
+        print(`   ${dbObj}`);
+      });
+
+      break;
+
+    case 'collections':
+      targetList.forEach(function (dbObj) {
+        var db = dbObj.db;
+        var cols = dbObj.cols;
+
+        cols.forEach(function (col) {
+          print(`   ${db}.${col}`);
+        });
+
+      });
+
+      break;
+
+    case 'indexes':
+      targetList.forEach(function(dbObj) {
+        var ns = dbObj.ns;
+        var idxs = dbObj.indexes;
+
+        idxs.forEach(function (idx) {
+          if (idx.name !== '_id_') {
+            print(`   ${ns}.${idx.name}`);
+          }
+        });
+      });
+
+      break;
+
+    default:
+      return false;
+  }
+
+  print('');
+  print(`Enter 'YES' (in all CAPS) at the 'Enter password' prompt to confirm this action`);
+
+  userInput = passwordPrompt();
+  print('')
+
+  return userInput === 'YES';
+}
+
+function silent() { return true; }
